refactor: split up Actions component and file
